Refactor HiveMetadata and Introduce MetadataUtils in presto-hive-common
